:star: add navbar navigation to dashboard
