+++ src/components/SigninForm/SigninForm.test.tsx
@@ -4,8 +4,6 @@ import renderer from 'react-test-renderer';
 import SigninForm from './SigninForm';
 
 it('renders SigninForm unchanged', () => {
-  const tree = renderer
-    .create(<SigninForm error={null} loading={false} onSubmit={jest.fn()} />)
-    .toJSON();
+  const tree = renderer.create(<SigninForm />).toJSON();
   expect(tree).toMatchSnapshot();
 });
