+++ src/components/SignupForm/SignupForm.test.tsx
@@ -4,8 +4,6 @@ import renderer from 'react-test-renderer';
 import SignupForm from './SignupForm';
 
 it('renders SignupForm unchanged', () => {
-  const tree = renderer
-    .create(<SignupForm error={null} loading={false} onSubmit={jest.fn()} />)
-    .toJSON();
+  const tree = renderer.create(<SignupForm />).toJSON();
   expect(tree).toMatchSnapshot();
 });
